+++ WebApplication/src/components/LabelsEditor.vue
@@ -11,7 +11,7 @@ window.filterLabel = (str) => {
 }
 
 export default {
-    props: ['labels', 'studyId', 'id', 'title'],
+    props: ['labels', 'studyId', 'id', 'title', 'showTitle', 'isFilter'],
     setup() {
     },
     data() {
@@ -22,13 +22,7 @@ export default {
         };
     },
     async created() {
-        if (this.canCreateNewLabels) {
-            // if we can create new labels, we already provide all the existing ones in the list
-            this.allLabelsLocalCopy = [...this.allLabels];
-        } else {
-            // if we can not create new labels, we only provide the available ones
-            this.allLabelsLocalCopy = [...this.uiOptions.AvailableLabels];
-        }
+        this.init();
         this.labelsModel = this.labels;
         // console.log("LabelsEditor: created ", this.labelsModel);
     },
@@ -39,7 +33,15 @@ export default {
         ...mapState({
             uiOptions: state => state.configuration.uiOptions,
             allLabels: state => state.labels.allLabels,
+            isLabelsLoaded: state => state.labels.loaded
         }),
+        isTitleVisible() {
+            if (this.showTitle == undefined) {
+                return true;
+            } else {
+                return this.showTitle;
+            }
+        },
         getTitle() {
             if (!this.title) {
                 return this.$t('labels.study_details_title');
@@ -47,9 +49,16 @@ export default {
             return this.$t(this.title)
         },
         canCreateNewLabels() {
+            if (this.isFilter == true) {
+                return false;
+            }
             return this.uiOptions.AvailableLabels.length == 0; // if there is no AvailableLabels list, this means we can create any new label
         },
         placeHolderText() {
+            if (this.isFilter == true) {
+                return this.$t('labels.filter_labels_placeholder')
+            }
+
             if (this.canCreateNewLabels) {
                 return this.$t('labels.add_labels_placeholder');
             } else {
@@ -63,8 +72,24 @@ export default {
             // console.log("hasLabel ", label, this.labelsModel.includes(label));
             return this.labelsModel.includes(label);
         },
+        init() {
+            // console.log("initializing LabelsEditor ", this.isFilter, this.allLabels, this.uiOptions.AvailableLabels);
+            if (this.isFilter) {
+                this.allLabelsLocalCopy = [...this.allLabels];
+            } else if (this.canCreateNewLabels && this.allLabels) {
+                // if we can create new labels, we already provide all the existing ones in the list
+                this.allLabelsLocalCopy = [...this.allLabels];
+            } else if (!this.canCreateNewLabels && this.uiOptions.AvailableLabels) {
+                // if we can not create new labels, we only provide the available ones
+                this.allLabelsLocalCopy = [...this.uiOptions.AvailableLabels];
+            }
+        }
     },
     watch: {
+        async isLabelsLoaded(newValue, oldValue) {
+            // console.log("labels loaded, init LabelsEditor")
+            this.init();            
+        },
         labelsModel: {
             async handler(newValue, oldValue) {
                 // console.log("LabelsEditor: labelsModel watcher", oldValue, newValue, this.disableWatcher);
@@ -95,12 +120,12 @@ export default {
 
 <template>
     <div :id="id">
-        <label :for="'select-' + id" class="form-label"><span v-html="getTitle"></span><span
+        <label v-if="showTitle" :for="'select-' + id" class="form-label"><span v-html="getTitle"></span><span
                 class="invalid-label-tips invalid-label-tips-hidden">{{ $t('labels.valid_alphabet_warning')
                 }}</span>
         </label>
 
-        <select class="form-select" :id="'select-' + id" name="tags[]" v-model="labelsModel" multiple
+        <select class="form-select" :id="'select-' + id" name="select-labels" v-model="labelsModel" multiple
             data-allow-clear="true" :data-allow-new="canCreateNewLabels" data-badge-style="info" data-input-filter="filterLabel"
             :placeholder="placeHolderText">
             <option v-for="label in allLabelsLocalCopy" :key="label" :value="label" :selected="hasLabel(label)">{{ label